requirements.txt
#git+https://github.com/everafactory/evera-messages@<branchname>
--extra-index-url https://builds.evera.network --trusted-host build.evera.network
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.22.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
Babel==2.6.0
base58==0.2.5
bitcoin==1.1.42
bitstring==3.1.5
cached-property==1.4.2
cbor==1.0.0
certifi==2018.1.18
cffi==1.10.0
chardet==3.0.4
click==6.7
coincurve==7.1.0
constantly==15.1.0
crossbar==17.12.1
cryptography==2.2.2
cytoolz==0.9.0.1
distro==1.3.0
dnspython==1.15.0
docker==3.5.0
enforce==0.3.4
enum34==1.1.6
eth-abi==1.1.1
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-tester==0.1.0-beta.24
eth-utils==1.0.3
ethereum==1.6.1
Evera-Messages==3.3.0
Evera-Smart-Contracts-Interface==1.6.1
greenlet==0.4.15
h2==3.0.1
hpack==3.0.0
html2text==2018.1.9
humanize==0.5.1
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
ipaddress==1.0.19
Jinja2==2.10
lmdb==0.94
MarkupSafe==1.0
miniupnpc==2.0.2
mistune==0.8.3
mock==2.0.0
monotonic==1.5
Naked==0.1.31
ndg-httpsclient==0.4.4
netaddr==0.7.19
netifaces==0.10.4
numpy==1.13.1
olefile==0.45.1
opencv-contrib-python-headless==3.4.3.18
OpenEXR==1.3.0
os-win==4.1.1
oslo.concurrency==3.28.1
oslo.config==6.5.1
oslo.context==2.21.0
oslo.i18n==3.22.1
oslo.log==3.40.1
oslo.serialization==2.28.1
oslo.utils==3.37.0
packaging==17.1
pathtools==0.1.2
pbkdf2==1.3
pbr==4.0.2
peewee-migrate==0.14.0
peewee==2.10.2
pluggy==0.6.0
portalocker==1.2.1
priority==1.3.0
psutil==5.2.2
py-cpuinfo==4.0.0
py-ubjson==0.11.0
pyasn1-modules==0.2.1
pyasn1==0.4.1
pycparser==2.17
pycryptodome==3.6.6
PyDispatcher==2.0.5
pyelliptic==1.5.10
pyethash==0.1.27
Pygments==2.2.0
pylru==1.0.9
PyNaCl==1.2.1
pyOpenSSL==17.1.0
pyparsing==2.2.0
PyQRCode==1.2.1
pysha3==1.0.2
pyssim==0.4
PyTrie==0.3
pytz==2018.4
PyYAML==3.12
raven==6.6.0
repoze.lru==0.7
requests==2.21.0
rlp==0.6.0
scrypt==0.8.6
sdnotify==0.3.2
secp256k1==0.13.2
semantic-version==2.6.0
service-identity==17.0.0
setproctitle==1.1.10
setuptools-rust==0.10.2
setuptools==38.7.0
shellescape==3.4.1
shutilwhich==1.1.0
six==1.11.0
sortedcontainers==1.5.9
tabulate==0.8.2
token-bucket==0.2.0
toolz==0.9.0
treq==17.8.0
Twisted==17.9.0
txaio==2.8.2
txtorcon==0.20.0
u-msgpack-python==2.5.0
urllib3==1.23
watchdog==0.8.3
web3==4.2.1
websocket-client==0.47.0
wrapt==1.10.11
zope.interface==4.4.2
zxcvbn-python==4.4.22
